requirements.txt
alembic==1.7.5
flasgger==0.9.5
Flask==1.1.1
Flask_Cors==3.0.10
Flask_Login==0.5.0
Flask_Migrate==2.7.0
Flask_Script==2.0.6
Flask_Session==0.4.0
Flask_SQLAlchemy==2.5.1
gevent==21.1.2
pycryptodomex==3.15.0
PyMySQL==1.0.2
redis==3.5.3
requests==2.22.0
SQLAlchemy==1.4.20
Werkzeug==0.16.0
gunicorn==20.0.4
